requirements.txt
fever-allennlp-reader
fever-scorer
prettytable
git+ssh://git@github.com/j6mes/fever-allennlp@master
sklearn
tqdm